backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client 
transformers
torch
scikit-learn 